requirements.txt
fastapi
uvicorn
numpy
tensorflow-cpu
num2words
gtts
TTS
pydantic
python-multipart
